size_t to int
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@20120 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/BzMaterial.cxx
@@ -76,7 +76,7 @@ const BzMaterial* BzMaterialManager::addMaterial(const BzMaterial* material)
   if (findMaterial(newMat->getName()) != NULL) {
     newMat->setName("");
   }
-  newMat->setID(materials.size());
+  newMat->setID((int)materials.size());
   materials.push_back(newMat);
   return newMat;
 }
